#79 Use TypeScript also in `gatsby-node.js`
