+++ package.json
@@ -26,6 +26,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-helmet": "^6.0.0",
+    "ts-node": "^8.10.1",
     "typescript": "^3.8.3"
   },
   "devDependencies": {
